% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REBASE_sequence_generator.R
\name{check_extended_compatible}
\alias{check_extended_compatible}
\title{Check if extended sequence is compatible with target sequence}
\usage{
check_extended_compatible(extended_seq, target_seq)
}
\description{
Check if extended sequence is compatible with target sequence
}
