% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REBASE_sequence_generator.R
\name{check_full_extended_match}
\alias{check_full_extended_match}
\title{Check if extended sequence fully matches target sequence}
\usage{
check_full_extended_match(extended_seq, target_seq)
}
\arguments{
\item{extended_seq}{A character string representing the extended sequence.}

\item{target_seq}{A character string representing the target sequence.}
}
\value{
A logical value (\code{TRUE} or \code{FALSE}) indicating whether the sequences match fully.
}
\description{
Compares an extended sequence with a target sequence, considering ambiguous nucleotide matching.
}
\examples{
check_full_extended_match("CCWGG", "CCAGG")

}
