% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REBASE_sequence_generator.R
\name{check_mod_position_in_overlap}
\alias{check_mod_position_in_overlap}
\title{Check if modification position is within the overlap region}
\usage{
check_mod_position_in_overlap(
  context,
  forward_RecSeq,
  mod_position,
  overlap_len
)
}
\arguments{
\item{context}{A character string representing the modification-containing sequence.}

\item{forward_RecSeq}{A character string representing the recognition sequence.}

\item{mod_position}{An integer representing the modification position in \code{context}.}

\item{overlap_len}{An integer indicating the length of the overlap.}
}
\value{
A logical value (\code{TRUE} or \code{FALSE}) indicating whether the modification position is within the overlap region.
}
\description{
Determines whether the modification position falls within the overlap region between two sequences.
}
\examples{
check_mod_position_in_overlap("CCWGG", "CCAGG", 2, 3)

}
