% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REBASE_sequence_generator.R
\name{concat_with_overlap}
\alias{concat_with_overlap}
\title{Concatenate two strings by removing duplicated ambiguous overlap}
\usage{
concat_with_overlap(a, b, mod_position)
}
\arguments{
\item{a}{Character string (mod_containing_seq).}

\item{b}{Character string (forward_RecSeq).}

\item{mod_position}{Integer. The 1-based position of the modification in \code{a}.}
}
\value{
A concatenated character string.
}
\description{
Concatenate two strings by removing duplicated ambiguous overlap
}
\examples{
concat_with_overlap("CGAAN", "GAAGA", 4)  # returns "CGAAGA"
}
