% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REBASE_sequence_generator.R
\name{cross_join_and_filter_na}
\alias{cross_join_and_filter_na}
\title{Cross join and filter NA values}
\usage{
cross_join_and_filter_na(df_filtered, df_meth_blocked)
}
\arguments{
\item{df_filtered}{A data frame containing sequence windows.}

\item{df_meth_blocked}{A data frame containing blocked enzyme data.}
}
\value{
A data frame with cross-joined data and removed NA values.
}
\description{
Performs a cross join between two data frames and removes rows with NA values in key columns.
}
\examples{
df_joined <- cross_join_and_filter_na(df_filtered, df_meth_blocked)

}
