% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REBASE_sequence_generator.R
\name{extend_to_target_forward}
\alias{extend_to_target_forward}
\title{Extend concatenated sequence to match target length (FORWARD direction)}
\usage{
extend_to_target_forward(concat_seq, target_seq)
}
\description{
Extend concatenated sequence to match target length (FORWARD direction)
}
