% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REBASE_sequence_generator.R
\name{extend_to_target_seq}
\alias{extend_to_target_seq}
\title{Extend concatenated sequence to target sequence}
\usage{
extend_to_target_seq(concatenated_RecSeq, target_seq)
}
\arguments{
\item{concatenated_RecSeq}{A character string representing the concatenated recognition sequence.}

\item{target_seq}{A character string representing the target DNA sequence.}
}
\value{
A character string representing the extended sequence or \code{NA} if no extension is possible.
}
\description{
Extends a concatenated recognition sequence to match the target sequence using ambiguous nucleotide matching.
}
\examples{
extend_to_target_seq("CCWGG", "CCAGG")

}
