% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REBASE_sequence_generator.R
\name{extend_with_target_seq_and_filter}
\alias{extend_with_target_seq_and_filter}
\title{Extend sequence to match target sequence and filter}
\usage{
extend_with_target_seq_and_filter(df_out, target_seq)
}
\arguments{
\item{df_out}{Data frame with concatenated sequences.}

\item{target_seq}{Character string representing the target DNA sequence.}
}
\value{
Filtered data frame with extended sequences.
}
\description{
Extend sequence to match target sequence and filter
}
\examples{
df_extended <- extend_with_target_seq_and_filter(df_out, "CCWGG")

}
