% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REBASE_methylsensitivity.R
\name{fetch_rebase_methylation}
\alias{fetch_rebase_methylation}
\title{Fetch and Process REBASE Methylation Sensitivity Data}
\usage{
fetch_rebase_methylation(
  save_to_file = FALSE,
  output_path = "REBASE_MethylSensitivity.xlsx"
)
}
\arguments{
\item{save_to_file}{Logical. If TRUE, saves the output to an Excel file. Default is FALSE.}

\item{output_path}{Character. Path to save the Excel file if save_to_file is TRUE. Default is "REBASE_MethylSensitivity.xlsx".}
}
\value{
A data frame containing processed methylation sensitivity data with columns for enzyme name,
recognition sequence, modification position, complementary modification position, cleavage information, and modification type.
}
\description{
This function fetches methylation sensitivity data for 4mC, 5mC, and 6mA modifications from REBASE,
processes the data into a tidy format, and optionally saves it as an Excel file.
}
\examples{
df <- fetch_rebase_methylation()
df_with_file <- fetch_rebase_methylation(save_to_file = TRUE, output_path = "methylation_data.xlsx")

}
