% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REBASE_sequence_generator.R
\name{find_intersection_in_target}
\alias{find_intersection_in_target}
\title{Find intersection between concatenated sequence and target sequence}
\usage{
find_intersection_in_target(concatenated_RecSeq, target_seq)
}
\arguments{
\item{concatenated_RecSeq}{A character string representing the concatenated recognition sequence.}

\item{target_seq}{A character string representing the target sequence.}
}
\value{
The longest matching substring or \code{NA} if no match is found.
}
\description{
Identifies the longest matching substring between a concatenated recognition sequence and a target sequence.
}
\examples{
find_intersection_in_target("CCWGG", "CCAGG")

}
