% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REBASE_sequence_generator.R
\name{longest_overlap}
\alias{longest_overlap}
\title{Find the longest overlap between two strings with ambiguous matching}
\usage{
longest_overlap(a, b, mod_position)
}
\arguments{
\item{a}{A character string (mod_containing_seq).}

\item{b}{A character string (forward_RecSeq).}

\item{mod_position}{Integer. The 1-based position of the modification in \code{a}.}
}
\value{
An integer representing the length of the longest overlap.
}
\description{
Find the longest overlap between two strings with ambiguous matching
}
\examples{
longest_overlap("CGAAN", "GAAGA", 4)  # returns 2 (AA와 GA 오버랩)
}
