% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REBASE_sequence_generator.R
\name{longest_overlap_backward}
\alias{longest_overlap_backward}
\title{Find longest BACKWARD overlap (RecSeq suffix matches window prefix)}
\usage{
longest_overlap_backward(window, recseq, mod_pos)
}
\description{
Find longest BACKWARD overlap (RecSeq suffix matches window prefix)
}
