% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REBASE_sequence_generator.R
\name{longest_overlap_forward}
\alias{longest_overlap_forward}
\title{Find longest FORWARD overlap (window suffix matches RecSeq prefix)}
\usage{
longest_overlap_forward(window, recseq, mod_pos)
}
\description{
Find longest FORWARD overlap (window suffix matches RecSeq prefix)
}
