% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REBASE_sequence_generator.R
\name{rev_comp}
\alias{rev_comp}
\title{Reverse complement of a DNA sequence}
\usage{
rev_comp(dna)
}
\arguments{
\item{dna}{A character string representing a DNA sequence.}
}
\value{
A character string representing the reverse complement of the input sequence.
}
\description{
Computes the reverse complement of a DNA sequence considering IUPAC nucleotide codes.
}
\examples{
rev_comp("ATGC") # returns "GCAT"

}
