% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REBASE_sequence_generator.R
\name{validate_mod_position}
\alias{validate_mod_position}
\title{Validate mod_position: user's methylated position must match REBASE mod_position}
\usage{
validate_mod_position(
  extended_seq,
  target_seq,
  recseq,
  user_mod_position,
  rebase_mod_position
)
}
\description{
Validate mod_position: user's methylated position must match REBASE mod_position
}
