% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REBASE_sequence_generator.R
\docType{data}
\name{ambiguous_dna_values}
\alias{ambiguous_dna_values}
\title{Ambiguous DNA values for nucleotide matching}
\format{
An object of class \code{list} of length 15.
}
\usage{
ambiguous_dna_values
}
\value{
A named list where each element is a string of acceptable bases.
}
\description{
A list mapping each nucleotide (or ambiguous IUPAC code) to its possible base representations.
}
\examples{
ambiguous_dna_values$N  # returns "ACGT"

}
\keyword{datasets}
