% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REBASE_sequence_generator.R
\docType{data}
\name{ambiguous_dna_values}
\alias{ambiguous_dna_values}
\title{Ambiguous DNA values for nucleotide matching.}
\format{
An object of class \code{list} of length 15.
}
\usage{
ambiguous_dna_values
}
\value{
A list where each element is a string of acceptable bases.
}
\description{
A list mapping each nucleotide (or ambiguous code) to the set of acceptable bases.
}
\examples{
ambiguous_dna_values$N  # "ACGT"
}
\keyword{datasets}
