% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REBASE_sequence_generator.R
\name{ambiguous_match}
\alias{ambiguous_match}
\title{Check ambiguous match between two sequences}
\usage{
ambiguous_match(seq1, seq2)
}
\description{
Check ambiguous match between two sequences
}
