% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REBASE_sequence_generator.R
\name{concat_with_overlap}
\alias{concat_with_overlap}
\title{Concatenate two strings by removing duplicated ambiguous overlap.}
\usage{
concat_with_overlap(a, b)
}
\arguments{
\item{a}{A character string.}

\item{b}{A character string.}
}
\value{
A concatenated character string.
}
\description{
Given two strings \code{a} and \code{b}, this function uses \code{longest_overlap()}
to determine the length of the overlapping region between the suffix of \code{a} and
the prefix of \code{b} (with ambiguous nucleotide matching). It then concatenates
\code{a} with the remainder of \code{b} after the overlap.
}
\examples{
concat_with_overlap("GAA", "AAGT")  # returns "GAAGT"

}
