% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REBASE_sequence_generator.R
\name{filter_meth_blocked}
\alias{filter_meth_blocked}
\title{Filter methylation sensitivity data for blocked enzymes}
\usage{
filter_meth_blocked(methylation_sensitivity)
}
\arguments{
\item{methylation_sensitivity}{A data frame from REBASE containing enzyme information.}
}
\value{
A filtered data frame with only blocked enzymes.
}
\description{
Filters methylation sensitivity data to retain only enzymes whose cleavage is blocked.
}
\examples{
df_blocked <- filter_meth_blocked(methylation_sensitivity)

}
