% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REBASE_sequence_generator.R
\name{filter_meth_by_mod_type}
\alias{filter_meth_by_mod_type}
\title{Filter methylation sensitivity data by modification type}
\usage{
filter_meth_by_mod_type(methylation_sensitivity, mod_type)
}
\description{
Filter methylation sensitivity data by modification type
}
