% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REBASE_sequence_generator.R
\name{longest_overlap}
\alias{longest_overlap}
\title{Find the longest overlap between two strings with ambiguous matching.}
\usage{
longest_overlap(a, b)
}
\arguments{
\item{a}{A character string.}

\item{b}{A character string.}
}
\value{
An integer representing the length of the longest overlap.
}
\description{
For two input strings \code{a} and \code{b}, this function finds the length of
the largest suffix of \code{a} that is also a prefix of \code{b}. Ambiguous nucleotides
are matched according to the values in \code{ambiguous_dna_values}. If either string is
\code{NA} or empty, the function returns 0.
}
\examples{
longest_overlap("GAA", "AAGT")  # returns 2 because "AA" overlaps.

}
