% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REBASE_sequence_generator.R
\name{match_enzyme_sequences}
\alias{match_enzyme_sequences}
\title{Match enzyme sequences with target sequence}
\usage{
match_enzyme_sequences(mod_type, target_seq)
}
\arguments{
\item{mod_type}{Character string (e.g., "4mC").}

\item{target_seq}{Character string (target DNA sequence).}
}
\value{
Data frame with matched enzyme sequences.
}
\description{
Match enzyme sequences with target sequence
}
