% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REBASE_sequence_generator.R
\name{rev_comp}
\alias{rev_comp}
\title{Reverse complement helper for uppercase A/T/G/C.}
\usage{
rev_comp(dna)
}
\arguments{
\item{dna}{A character string representing a DNA sequence.}
}
\value{
The reverse complement of \code{dna}.
}
\description{
Reverse complement helper for uppercase A/T/G/C.
}
\examples{
rev_comp("ATGC")  # returns "GCAT"

}
